+++ ChangeLog.md
@@ -4,7 +4,7 @@ Sessions for the XP Framework ChangeLog
 ## 0.6.0 / ????-??-??
 
 * Changed session cookie to be transmitted via HTTPS only by default.
-* Added web.session.Session::disableSecure() to restore old behaviour.
+* Added web.session.Session::insecure(bool $whether) to restore old behaviour.
 
 ## 0.5.0 / 2018-08-22